refactor : 투표 상태 및 투표 실시간 업데이트 로직을 useBattleSocket 훅으로 이동
- battlePage/index.tsx에서 관리하는 투표 상태 및 실시간 업데이트 로직을 hook에서 관리하도록 로직 수정
- 현재 상태 관련 로직이 battlePage/index.tsx에 많은 로직들이 집중되어있는데 점차 hook들로 분리 예정
